:bug: FT_68: Payment Term Settings Migration - Fixed **code** column must not be unique in database

+++ database/migrations/2022_06_09_105925_remove_foreign_key_constraints_in_cdis_terminal_transaction_discount_table.php
@@ -1,11 +1,13 @@
 <?php
 
+use App\Traits\MigrationTrait;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
 class RemoveForeignKeyConstraintsInCdisTerminalTransactionDiscountTable extends Migration
 {
+    use MigrationTrait;
     /**
      * Run the migrations.
      *
@@ -15,7 +17,9 @@ class RemoveForeignKeyConstraintsInCdisTerminalTransactionDiscountTable extends 
     {
         Schema::disableForeignKeyConstraints();
         Schema::table('cdis_terminal_transaction_discount', function (Blueprint $table) {
-            $table->dropForeign('discount_transaction_product_bid_foreign');
+            if ($this->hasForeignKey('cdis_terminal_transaction_discount', 'discount_transaction_product_bid_foreign')) {
+                $table->dropForeign('discount_transaction_product_bid_foreign');
+            }
         });
         Schema::enableForeignKeyConstraints();
     }
@@ -28,11 +32,13 @@ class RemoveForeignKeyConstraintsInCdisTerminalTransactionDiscountTable extends 
     public function down()
     {
         Schema::table('cdis_terminal_transaction_discount', function (Blueprint $table) {
-            $table->foreign('transaction_product_bid', 'discount_transaction_product_bid_foreign')
-            ->references('bid')
-            ->on('terminal_transaction_product')
-            ->onUpdate('restrict')
-            ->onDelete('cascade');
+            if (! $this->hasForeignKey('cdis_terminal_transaction_discount', 'discount_transaction_product_bid_foreign')) {
+                $table->foreign('transaction_product_bid', 'discount_transaction_product_bid_foreign')
+                    ->references('bid')
+                    ->on('cdis_terminal_transaction_product')
+                    ->onUpdate('restrict')
+                    ->onDelete('cascade');
+            }
         });
     }
 }
